part of bug 4720
git-svn-id: svn+ssh://svn.abisource.com/svnroot/abiword/trunk@10921 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6

+++ src/af/xap/unix/xap_UnixFrameImpl.cpp
@@ -1209,7 +1209,7 @@ gint XAP_UnixFrameImpl::_imDeleteSurrounding_cb (GtkIMContext *slave, gint offse
 	FV_View * pView = static_cast<FV_View*>(pImpl->getFrame()->getCurrentView ());
 
 	PT_DocPosition insPt = pView->getInsPoint ();
-	if (offset > insPt)
+	if (offset > (gint)insPt)
 		insPt = 0;
 	else
 		insPt += offset;
